Fixed a few small bugs
The search result's endless scrolling was never unbinded/removed. Fixed a bit of the code to use .get() over .attributes, makes it a bit more readable. In the course listing of sections, if couchdb is down, but node.js still up the code produces errors.  I've now fixed it so that if node.js returns an empty array [] (aka no grades), it will not crash the webapp.