Minor - Fixed a styling bug
In the search results, the right justified table (with the gpa, stddev,...etc) and right justified buttons (sections/grades) was overlapping the course description text.  And it tended to bloat up the page a bit much on smaller screens. So I've changed this right justified area to only show on >= large screens.