No longer showing the schedule view for 'medium' screens
The schedule view is cut off for medium screens, so I've made it only visible on large screens.  Also did a minor html change for the advanced search.  Basically just changed to from saying advanced to a down arrow. Also added the search class to the advanced search modal's query input, so that it gets auto-populated with the already present search query